Job_Service/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
asyncpg
python-dotenv
icecream
pyjwt
sqlalchemy
httpx